detector_sellos/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
torch
torchvision
git+https://github.com/openai/CLIP.git
scikit-learn==1.3.2
cairosvg==2.7.1
python-multipart==0.0.6